requirements.txt
bitarray==1.2.2
cytoolz==0.11.2
discord==1.7.3
discord.py==1.7.3
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
jsonschema==3.2.0
lru-dict==1.1.7
python-dotenv==0.19.2
requests==2.27.1
web3==5.26.0